+++ Packs/AutoFocus/Integrations/AutofocusV2/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-Better error message for connection errors
+Added support for improved error messages for server connection issues.
 
 
 ## [20.4.1] - 2020-04-29